@@ -3,40 +3,50 @@
   "version": "1.0.0",
   "description": "Starter point for a graphQL course",
   "main": "index.js",
+  "type": "module",
   "repository": {
     "type": "git",
     "url": "https://github.com/StephenGrider/Lyrical-GraphQL"
   },
   "scripts": {
-    "dev": "nodemon index.js --ignore client"
+    "start": "nodemon index.js --ignore client"
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "apollo-client": "^0.8.1",
+    "@apollo/client": "^3.7.7",
+    "@chakra-ui/react": "^1.8.9",
+    "@emotion/react": "^11.10.6",
+    "@emotion/styled": "^11.10.6",
     "axios": "^0.15.3",
-    "babel-core": "^6.22.1",
-    "babel-loader": "^6.2.10",
-    "babel-preset-env": "^1.1.8",
-    "babel-preset-react": "^6.22.0",
     "body-parser": "^1.16.0",
     "connect-mongo": "^1.3.2",
     "css-loader": "^0.26.1",
     "express": "^4.14.0",
-    "express-graphql": "0.6.1",
+    "express-graphql": "0.12.0",
     "express-session": "^1.15.0",
-    "graphql": "^0.8.2",
-    "html-webpack-plugin": "^2.26.0",
+    "framer-motion": "4.1.17",
+    "graphql": "^14.3.1",
+    "graphql-tag": "^2.12.6",
+    "html-webpack-plugin": "^4.5.2",
     "lodash": "^4.17.4",
-    "mongoose": "^4.7.8",
+    "mongoose": "^6.9.1",
+    "nodemon": "^2.0.20",
     "passport": "^0.3.2",
     "passport-local": "^1.0.0",
-    "react": "^15.4.2",
-    "react-apollo": "^0.9.0",
-    "react-dom": "^15.4.2",
-    "react-router": "^3.0.2",
-    "style-loader": "^0.13.1",
-    "webpack": "^2.2.0",
-    "webpack-dev-middleware": "^1.9.0"
+    "react": "^16.8.0",
+    "react-apollo": "^3.1.5",
+    "react-dom": "^16.8.0",
+    "react-router": "^6.8.1",
+    "react-router-dom": "^6.8.1",
+    "style-loader": "^2.0.0",
+    "webpack": "^4.46.0",
+    "webpack-dev-middleware": "^5.0.0"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.20.12",
+    "@babel/preset-env": "^7.20.2",
+    "@babel/preset-react": "^7.18.6",
+    "babel-loader": "^8.3.0"
   }
 }
