@@ -1,4 +1,4 @@
-const app = require('./server/server');
+import app from './server/server.js';
 
 app.listen(4000, () => {
   console.log('Listening');
