@@ -1,5 +1,5 @@
 const _ = require('lodash');
-const graphql = require('graphql');
+const graphql = require('graphql')
 const { GraphQLSchema } = graphql;
 
 const RootQueryType = require('./root_query_type');
