@@ -1,5 +1,5 @@
 const mongoose = require('mongoose');
-const graphql = require('graphql');
+const graphql = require('graphql')
 const { GraphQLObjectType, GraphQLString, GraphQLID, GraphQLList } = graphql;
 const LyricType = require('./lyric_type');
 const Song = mongoose.model('song');
