@@ -2,27 +2,36 @@ const webpack = require('webpack');
 const HtmlWebpackPlugin = require('html-webpack-plugin');
 
 module.exports = {
-  entry: './client/index.js',
-  output: {
-    path: '/',
-    filename: 'bundle.js'
-  },
-  module: {
-    rules: [
-      {
-        use: 'babel-loader',
-        test: /\.js$/,
-        exclude: /node_modules/
-      },
-      {
-        use: ['style-loader', 'css-loader'],
-        test: /\.css$/
-      }
+    entry: './client/index.js',
+    output: {
+        path: '/',
+        filename: 'bundle.js'
+    },
+    mode: 'development',
+    module: {
+        rules: [
+            {
+                test: /\.js$/,
+                exclude: /node_modules/,
+                use: {
+                    loader: 'babel-loader',
+                    options: {
+                        presets: ['@babel/preset-react'],
+                        // plugins: [
+                        //   ["@babel/plugin-proposal-decorators", { "legacy": true }],
+                        // ]
+                    }
+                }
+            },
+            {
+                test: /\.css$/,
+                use: ['style-loader', 'css-loader']
+            }
+        ]
+    },
+    plugins: [
+        new HtmlWebpackPlugin({
+            template: 'client/index.html'
+        })
     ]
-  },
-  plugins: [
-    new HtmlWebpackPlugin({
-      template: 'client/index.html'
-    })
-  ]
 };
